feat!: adapt Light.DataAccessMocks to Light.SharedCore v3
