+++ Light.DataAccessMocks/Light.DataAccessMocks.csproj
@@ -33,7 +33,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Light.SharedCore" Version="2.0.0" />
+    <PackageReference Include="Light.SharedCore" Version="3.0.0" />
     <PackageReference Include="Humanizer.Core" Version="2.14.1" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
   </ItemGroup>
